requirements.txt
streamlit
requests
pinecone-client
transformers
pandas
numpy
torch
newsapi-python
python-dotenv